// product-service/src/common/config/custom-config.enum.ts
export enum EnvKey {
  NODE_ENV = 'NODE_ENV',
  NODE_PORT = 'NODE_PORT',

  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_PORT = 'DATABASE_PORT',
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_PASSWORD = 'DATABASE_PASSWORD',
  DATABASE_NAME = 'DATABASE_NAME',

  RABBIT_MQ_URI = 'RABBIT_MQ_URI',
  RABBIT_MQ_TIMEOUT = 'RABBIT_MQ_TIMEOUT',

  AUTH_COOKIE_NAME = 'AUTH_COOKIE_NAME',
  AUTH_COOKIE_DOMAIN = 'AUTH_COOKIE_DOMAIN',

  LANDING_URL = 'LANDING_URL',

  ALLOWED_ORIGINS = 'ALLOWED_ORIGINS',
}